# [SOM] Fix overlapping bounding boxes and added GPU/MPS support