- Added GPU and MPS support for easyOCR to improve performance
- Fixed timing miscalculation in SOM experiment example
- Implemented a workaround for handling large overlapping bounding boxes (#78)
- Applied non-maximum suppression (NMS) merge to OCR detections to reduce duplicates (#78)
- Included detection bounding boxes in the label collision check process (#78)

| Before: | After: |
| -------- | ------ |
| ![image](https://github.com/user-attachments/assets/a168252c-412d-4a85-bbbc-4de137dc1c71) | ![image](https://github.com/user-attachments/assets/cc822c4c-e194-46ff-bfd8-193e082a3864) |
| ~3.46s | ~1.63s |